requirements.txt
Flask==3.0.0
flask-mysqldb==2.0.0
mysqlclient==2.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0